Major EJB overhaul:
 * Adds RawPixelsBean (#170)
 * Fixes RenderingBean (#174) (see SessionHandler)
 * Fixes passivation (#173)
 * Reworks the Spring config (in-/out-of-container.xml)
 * Renames methods in ServiceFactory



git-svn-id: file:///home/svn/omero/trunk@731 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/server/src/ome/logic/PojosImpl.java
@@ -542,7 +542,7 @@ public class PojosImpl extends AbstractLevel2Service implements IPojos
                 for (Object[] results : l_c)
                 {
                    Long id = (Long) results[0];
-                   Integer count = (Integer) results[1];
+                   Long count = (Long) results[1];
                    c.addCounts(key,id,count);
                 }
